backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
openai==1.50.0
faiss-cpu==1.8.0
boto3==1.35.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
pydantic==2.9.0
PyPDF2==3.0.1
tiktoken==0.7.0
numpy==1.26.4
python-multipart==0.0.9
packaging==26.0
beautifulsoup4==4.14.3
requests==2.32.5
lxml==6.0.2
